refactor: MemberService와 AuthService 분리
