+++ src/main/java/com/server/crews/auth/presentation/AuthController.java
@@ -1,10 +1,10 @@
 package com.server.crews.auth.presentation;
 
 import com.server.crews.auth.application.AuthService;
+import com.server.crews.auth.application.MemberService;
 import com.server.crews.auth.domain.Role;
 import com.server.crews.auth.dto.request.AdminLoginRequest;
 import com.server.crews.auth.dto.request.ApplicantLoginRequest;
-import com.server.crews.auth.dto.request.NewApplicantRequest;
 import com.server.crews.auth.dto.response.AccessTokenResponse;
 import com.server.crews.auth.dto.response.RefreshTokenDto;
 import io.swagger.v3.oas.annotations.Operation;
@@ -24,11 +24,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class AuthController {
     private final AuthService authService;
+    private final MemberService memberService;
 
     @PostMapping("/recruitment/login")
     @Operation(description = "[동아리 관리자] 로그인 해 토큰을 발급 받는다. 모집 공고가 존재하지 않는다면 모집 공고를 새로 생성한다.")
     public ResponseEntity<AccessTokenResponse> createRecruitmentSecretCode(@RequestBody AdminLoginRequest request) {
-        AccessTokenResponse accessTokenResponse = authService.loginForAdmin(request);
+        AccessTokenResponse accessTokenResponse = memberService.loginForAdmin(request);
         RefreshTokenDto refreshTokenDto = authService.createRefreshToken(Role.ADMIN, accessTokenResponse.memberId());
         ResponseCookie cookie = refreshTokenDto.toCookie();
         return ResponseEntity.status(HttpStatus.CREATED)
@@ -39,7 +40,7 @@ public class AuthController {
     @PostMapping("/applicant/login")
     @Operation(description = "[지원자] 로그인 해 토큰을 발급 받는다.")
     public ResponseEntity<AccessTokenResponse> createApplicationSecretCode(@RequestBody ApplicantLoginRequest request) {
-        AccessTokenResponse accessTokenResponse = authService.loginForApplicant(request);
+        AccessTokenResponse accessTokenResponse = memberService.loginForApplicant(request);
         RefreshTokenDto refreshTokenDto = authService.createRefreshToken(Role.APPLICANT, accessTokenResponse.memberId());
         ResponseCookie cookie = refreshTokenDto.toCookie();
         return ResponseEntity.status(HttpStatus.CREATED)
